requirements.txt
Django==1.8.2
djangorestframework==3.1.3
celery==3.1.18
django-celery==3.1.16
django-haystack==2.4.0
pyelasticsearch==1.4
requests==2.7.0